+++ lib/micro/case/with_test_doubles.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'micro/case'
+require 'micro/case/success'
+require 'micro/case/failure'